kernel-6.12: include efa from efa-installer
The EFA driver in the Amazon Linux kernel can have some delay between
when the installer is released and it making it into the kernel tree.
This adds a build step to update the driver from the installer to get
the latest changes as soon as the installer is available.

Signed-off-by: Matthew Yeazel <yeazelm@amazon.com>